# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API to interface with Sidero Omni.",
        "title": "Talos Omni Control API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/clustermachines": {
            "get": {
                "description": "Get a list of all cluster machines in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustermachines"
                ],
                "summary": "List all cluster machines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by cluster ID",
                        "name": "cluster",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ClusterMachineResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clustermachines/{id}/config": {
            "get": {
                "description": "Get machine configuration for a specific cluster machine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustermachines"
                ],
                "summary": "Get cluster machine config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ClusterMachineConfigResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clusters": {
            "get": {
                "description": "Get a list of all clusters in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "List all clusters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ClusterResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clusters/{id}/destroy-status": {
            "get": {
                "description": "Get the status of cluster destruction operation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "Get cluster destroy status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ClusterDestroyStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clusters/{id}/kubernetes-nodes": {
            "get": {
                "description": "Get a list of all Kubernetes nodes in a cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "List cluster Kubernetes nodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ClusterKubernetesNodeResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clusters/{id}/kubernetes-nodes/{node}": {
            "get": {
                "description": "Get detailed information about a specific Kubernetes node in a cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "Get a cluster Kubernetes node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node name",
                        "name": "node",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ClusterKubernetesNodeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ClusterDestroyStatusResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ClusterKubernetesNodeResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "handlers.ClusterMachineConfigResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "cluster_machine_version": {
                    "type": "string"
                },
                "config_size": {
                    "type": "integer"
                },
                "generation_error": {
                    "type": "string"
                },
                "grub_use_uki_cmdline": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "without_comments": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.ClusterMachineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "kubernetes_version": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "machine_id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ClusterResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "features": {
                    "type": "object",
                    "properties": {
                        "disk_encryption": {
                            "type": "boolean"
                        },
                        "workload_proxy": {
                            "type": "boolean"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "kubernetes_version": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                }
            }
        },
...
    }
}